+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Add a subcommand to remove a recipe or an image
 - Add a way to specify multiple versions for a recipe
 - Add a `versions` field to `cmd` to specify on which versions of recipe to run a command
+- Add a way to specify a base recipe with `from: base-recipe`. More information in the docs
 
 # 0.8.0
 - Dependencies defined for default images like `pkger-deb` or `pkger-rpm` will be used for custom images using the same output target. This means that all images used for building RPM will use dependencies defined for `pkger-rpm` - cleaner recipes ;)