+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+#
+- Add support for env variables in git source url
+
 # 0.9.0
 - Rename `--docker-uri` option to `--runtime-uri` and make it global (no longer part of build and init command)
 - Rename `docker_uri` configuration option to `runtime_uri`